requirements.txt
langchain-community
langchain-anthropic
langchain
langchain-huggingface
langchain-google-genai
langchain-openai
python-dotenv
mysql-connector-python
magic-wormhole
sycamore-ai
Flask
Flask-Cors
pandas
